Provide PoC for supplying mu-auth-allowed-groups header on callback
When we send a delta request to a consumer, we should inform the
consumer about the access rights used to create that change.  This may
help during invalidation (eg: to receive the type of the changed
instance) or during updating (to update related content in name of the
user).

The current implementation assumes all delta messages in one unit are
sent with the same access rights.  That is a simplification we should
not count on for the future.