Update genesis format and introduce v0.0.1
It seems we have misread the genesis format.  We have created a new
format to abide its constraints but don't have all the necessary
information to make that work.  With the new system, it is not as
clear in which specific graphs the content was changed as we now work
with access rights.  For this reason we can not fully mock the
v0.0.0-genesis format.

The format which we previously assumed was genesis, has been renamed
to be the v0.0.1 format.  Services working with these initial drafts
will need to be updated.