Provide a call id trail for messages received through Delta
When messages have been received through a Delta, there is now a call
id trail.  This trail helps identify the source of a certain
operation.  This is important as the Delta is used more often to cause
linked behaviour, thereby making the system less obvious to inspect
from a glance at the docker-compose file.